+++ experiments/Retrosynthesis/RetrosynthesisExperiment.py
@@ -1,4 +1,3 @@
-from typing import Optional
 from charge.Experiment import Experiment
 
 
@@ -23,16 +22,11 @@ class RetrosynthesisExperiment(Experiment):
     def __init__(
         self,
         user_prompt,
-        system_prompt: Optional[str] = None,  # Add optional parameter
     ):
-        # Use provided system prompt or fall back to default
-        if system_prompt is None:
-            system_prompt = SYSTEM_PROMPT
-        
         super().__init__(
-            system_prompt=system_prompt,
+            system_prompt=SYSTEM_PROMPT,
             user_prompt=user_prompt,
         )
         print("RetrosynthesisExperiment initialized with the provided prompts.")
-        self.system_prompt = system_prompt
+        self.system_prompt = SYSTEM_PROMPT
         self.user_prompt = user_prompt
